📖 DOC: Added docs to Config
